requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
future==0.17.1
gast==0.3.1
geojson==2.5.0
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
imutils==0.5.3
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.3
nltk==3.2.4
numpy==1.17.2
object-detection==0.1
opencv-python==4.1.1.26
pandas==0.25.1
Pillow==6.1.0
preprocessing==0.1.13
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
roadai-pkg==0.0.1
scikit-image==0.15.0
scikit-learn==0.21.3
scikit-video==1.1.11
scipy==1.3.1
six==1.12.0
sphinx-rtd-theme==0.2.4
tensorboard==1.14.0
#tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tinydb==3.14.1
tuspy==0.2.4
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
tensorflow-gpu==1.15
git+https://gitlab.com/vionice/roadai-python.git
Keras==2.2.4
scikit-plot==0.3.7
